Move the nested collate function into a separate callable object.
The original nested functions `collate` and `dense_collate` cause problems on Windows because subprocess (on Win) uses pickle for serialization when multiprocessing. Lambda and nested functions cannot be pickled on Windows. Therefore, `dense_collate` is moved out of `DenseDataLoader`  and `collate` is turned into a callable object `Collater` because it needs a member.