refactor(sampler): clean up sampler attributes, part 3 (#5402)
This PR continues the effort to consolidate PyG's sampling interface in preparation for moving `sample(...)` behind the `GraphStore` interface. This effort is somewhat large in scope and will be broken into multiple PRs for ease of review. It resolves some TODOs from #5365.

The major change removes special handling of `input_type` and `perm` / `perm_dict` in the neighbor loader. This enables a full separation between the sampler interface and the NeighborLoader (the only other interaction that is not clearly documented or exposed well in the sampler class is `SamplerOutput.metadata`, which will take a bit more work to properly define.