Add `HydroNet` dataset (#5537)
This PR adds support for working with the `HydroNet` dataset described
in:
[HydroNet: Benchmark Tasks for Preserving Intermolecular Interactions
and Structural Motifs in Predictive and Generative Models for Molecular
Data](https://arxiv.org/abs/2012.00131)

This is quite a large dataset of ~5 million graphs where each graph
consists of:

* `[num_atoms, 3]` tensor containing atomic coordinates
* `y` float label containing the energy for the water cluster in
kcal/mol

The raw dataset is stored as several compressed `.xyz` files where each
file contains the structures for the a fixed number of clusters (e.g.
the file `WN_geoms_all.zip` contains the data for a cluster of `N` water
molecules). The entire dataset is pre-processed in parallel using pandas
to read the compressed files into the PyG `Data` objects. The number of
workers for this is controlled by the `max_num_workers` argument. A
subset of the entire dataset can be selected through the `clusters`
argument of the constructor.

Some additional ideas (could be handled in either this PR or as a follow
up):
* a complete example using the SchNet implementation in PyG to train for
a few epochs on a subset of the dataset
* automated download from the official PNNL data hub (expect this will
need help from @jenna1701, @jesunsahariar, and @sutanay

Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>