Some usability updates to `Explanation` (#6054)
### Motivation
When evaluating and visualizing explanations it will be very convenient
to obtain the explanation and explanation complement subgraphs directly
from the `Explanation` objects. Given that the class holds information
of the original graph being explained this shouldn’t be difficult to
achieve. I propose we add two methods to the `Explanation` class:

1. `get_explanation_subgraph` returns the explanation subgraph $G_S$,
2. `get_complement_subgraph` returns the complement of the explanation
subgraph $G_{C \backslash S}$

I’ve submitted this draft PR to get opinions on this, both
implementation-wise and if we should do this in the first place.

### Alternatives
1. Delegate this to other parts of code, I.e. whenewer this is needed
(e.g. when evaluation explanations or visualizing them)
2. Implement the logic in `util.py` for general Data object, e.g. a
`mask_graph` method, which can then be reused here or elsewhere.

Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>