Drop `target_index` argument in `Explainer` (#6270)
I feel this is a very special argument that over-complicates a lot of
design decisions in the `Explainer` code, e.g., it needs to be passed to
`get_prediction` or `get_masked_prediction` as well. I am in favor of
dropping this argument. If the user wants to explain a model with
multiple outputs, she can use hooks to only return the desired output.