Add the unsupervised bipartite GraphSAGE model on the Taobao dataset (#6144)
This PR adds an implementation of unsupervised bipartite GraphSAGE on
the Taobao User Behaviors dataset offered by Alibaba. The Taobao dataset
contains a heterogeneous graph, where nodes represent users and items,
and edges represent different types of behaviors between users and
items. [](https://tianchi.aliyun.com/dataset/649)
We use the i2i co-occurrence matrix to construct the i2i-graph. When
applying GraphSAGE, the model follows the i-i-i pattern to encode the
item embedding and the i-i-u pattern to encode the user embedding. As
the task is unsupervised, we use negative sampling and
`binary_cross_entropy_with_logits` to compute the loss in the model.

Co-authored-by: huxleyhu <shuxian.hu98@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>