Huggingface hub integration 5545 (#5930)
This is an initial draft of what the wrapper could look like. 

Some questions I have:
- is this a sensible approach? (I have overridden the from_pretrained
and save_pretrained methods so we can have defaults for the model cards,
and a slightly different way of initialising the PyG model as some of
them require the data at init time)
- should the wrapper live somewhere else in the codebase? 
- do we want to have a nicer template for PyG models or just use the
default one? Are there fields we want to have prepopulated by default
- should the wrapper also contain methods to log the training/evaluation
metrics by default?
@rusty1s would be great if you comment :)

The screenshot is an example of what it currently looks like if you use
the wrapper to upload to huggingface model hub
![Screenshot from 2022-11-08
22-08-56](https://user-images.githubusercontent.com/6398157/200686722-a29c60a2-4f97-4268-b39a-97b4430089aa.png)

---------

Co-authored-by: Jinu Sunil <jinu.sunil@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>