Update `LabelPropagation` docs and `post_step` default value (#6749)
The typing of the arguments for the `LabelPropagation` `forward`
function is currently incorrect due to the use of a default lambda
function. To address this, changing the default value to `None`, and
moving the lambda function into the function itself. Additionally,
adding documentation explain the `post_step` for the output.

<img width="757" alt="image"
src="https://user-images.githubusercontent.com/21955420/219884703-766437e3-f9b0-428f-a411-2a42eb603ab1.png">

---------

Co-authored-by: Matthias Fey <matthias.fey@tu-dortmund.de>