`Explain`: Update `CaptumExplainer` and associated tests (#6910)
Currently `CaptumExplainer` only work for some `attribution_method`s. 
This PR lists and adds tests for **some** methods that work.

TODO in follow up PRs: support more attribution methods. Currently the
unsupported methods are methods that require a `baseline` argument and
some methods don't work when `index` is a tensor. Will open an issue for
this.

---------

Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>