Fix handling of missing node types in heterogeneous data loaders (#7060)
This addresses a bug in heterogeneous data loaders, which crash if a
node type from the complete graph is not present in the sampled batch of
nodes. This is can easily happen if a graph is disconnected or some node
types are simply rare.

An unsafe dictionary access in the `filter_hetero_data` raises an
exception in this case. I solved this by calling
`HeteroData.node_type_subgraph` on the graph before the batch of nodes
is filtered. I added a test for `HGTLoader` which covers this case and
did not pass previously.

Note that the behavior of the data loaders now differs slightly from
`HeteroData.subgraph`, as the loader will handle missing keys by not
including any node of that type in the batch, while the `subgraph`
method would retain all of the nodes. However, it seems more convenient
for data loading if batches only contain nodes that where actually
returned by the underlying sampler.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jinu Sunil <jinu.sunil@gmail.com>