Hypergraph data object (#7611)
Hello. I don't know the definition of a subgraph for a hypergraph (not
mentioned in papers). When getting a subgraph, a hyperedge can be cut.
For example, if a hyperedge connects vertices [1, 2, 3], but only [1, 3]
are selected for the subgraph, then we have the option to: 1) remove
this hyperedge; 2) change the hyperedge and make it only connects [1,
3]. So, in this version, I choose the second option.

---------

Co-authored-by: wsad1 <jinu.sunil@gmail.com>